RAG_Workbench/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
sentence-transformers
chromadb
langchain-ibm
# Add other dependencies from ibm_webinar.py as needed
